requirements.txt
selenium
pytest
allure-pytest
webdriver-manager